Installer shouldn't be adding bundled Ruby to PATH
As stated by the installer, The Ruby VM bundled with the installer
is isolated and for the unique purposes of running 'heroku' command
line and nothing else.

Adding it to the PATH affects other Ruby installations and not only
that, it was done at machine level (HKLM) affecting other user accounts on
the same computer.

Such modification is bad and produced several issues to other projects
like RailsInstaller, RubyStack and RubyInstaller users directly.